requirements.txt
langchain
langchain-community
langchain-core
groq
python-dotenv
pandas
faiss-cpu
langchain-openai
langchain-groq
